+++ cogs/plugins.py
@@ -3,12 +3,12 @@ import importlib
 import json
 import logging
 import os
-import random
 import shutil
 import site
 import stat
 import subprocess
 import sys
+import typing
 from difflib import get_close_matches
 
 import discord
@@ -57,7 +57,7 @@ class Plugins(commands.Cog):
         # default branch = master
         try:
             # when names are formatted with inline code
-            result = name.strip("`").split("/")
+            result = name.split("/")
             result[2] = "/".join(result[2:])
             if "@" in result[2]:
                 # branch is specified
@@ -83,13 +83,13 @@ class Plugins(commands.Cog):
                 await self.download_plugin_repo(username, repo, branch)
             except DownloadError as exc:
                 msg = f"{username}/{repo}@{branch} - {exc}"
-                logger.error(msg)
+                logger.error(msg, exc_info=True)
             else:
                 try:
                     await self.load_plugin(username, repo, name, branch)
                 except DownloadError as exc:
                     msg = f"{username}/{repo}@{branch}[{name}] - {exc}"
-                    logger.error(msg)
+                    logger.error(msg, exc_info=True)
 
     async def download_plugin_repo(self, username, repo, branch):
         try:
@@ -136,7 +136,7 @@ class Plugins(commands.Cog):
                     # so there's no terminal output unless there's an error
             except subprocess.CalledProcessError as exc:
                 err = exc.stderr.decode("utf8").strip()
-
+                logger.error('Error.', exc_info=True)
                 if err:
                     msg = f"Requirements Download Error: {username}/{repo}@{branch}[{plugin_name}]"
                     logger.error(msg)
@@ -154,7 +154,7 @@ class Plugins(commands.Cog):
             self.bot.load_extension(ext)
         except commands.ExtensionError as exc:
             msg = f"Plugin Load Failure: {username}/{repo}@{branch}[{plugin_name}]"
-            logger.error(msg)
+            logger.error(msg, exc_info=True)
             raise DownloadError("Invalid plugin") from exc
         else:
             msg = f"Loaded Plugin: {username}/{repo}@{branch}[{plugin_name}]"
@@ -174,9 +174,8 @@ class Plugins(commands.Cog):
 
         if plugin_name in self.registry:
             details = self.registry[plugin_name]
-            plugin_name = (
-                details["repository"] + "/" + plugin_name + "@" + details["branch"]
-            )
+            plugin_name = details["repository"] + "/" + plugin_name + "@" + details["branch"]
+
             required_version = details["bot_version"]
 
             if parse_version(self.bot.version) < parse_version(required_version):
@@ -212,7 +211,9 @@ class Plugins(commands.Cog):
 
                 try:
                     await self.download_plugin_repo(username, repo, branch)
-                except DownloadError as exc:
+                except Exception as exc:
+                    if not isinstance(exc, DownloadError):
+                        logger.error('Unknown error when adding a plugin:', exc_info=True)
                     embed = discord.Embed(
                         description=f"Unable to fetch this plugin from Github: `{exc}`.",
                         color=self.bot.main_color,
@@ -224,6 +225,8 @@ class Plugins(commands.Cog):
                 try:
                     await self.load_plugin(username, repo, name, branch)
                 except Exception as exc:
+                    if not isinstance(exc, DownloadError):
+                        logger.error('Unknown error when adding a plugin:', exc_info=True)
                     embed = discord.Embed(
                         description=f"Unable to load this plugin: `{exc}`.",
                         color=self.bot.main_color,
@@ -238,13 +241,15 @@ class Plugins(commands.Cog):
 
                 embed = discord.Embed(
                     description="The plugin is installed.\n"
-                    "*Please note: Any plugin that you install is at your **own risk***",
+                    "*Friendly reminder, plugins have absolute control over your bot. "
+                    "Please only install plugins from developers you trust.*",
                     color=self.bot.main_color,
                 )
                 await ctx.send(embed=embed)
             else:
                 embed = discord.Embed(
-                    description="Invalid plugin name format: use username/repo/plugin.",
+                    description="Invalid plugin name format: use plugin-name or "
+                                "username/repo/plugin or username/repo/plugin@branch.",
                     color=self.bot.main_color,
                 )
                 await ctx.send(embed=embed)
@@ -261,19 +266,17 @@ class Plugins(commands.Cog):
             )
 
         if plugin_name in self.bot.config["plugins"]:
+            username, repo, name, branch = self.parse_plugin(plugin_name)
             try:
-                username, repo, name, branch = self.parse_plugin(plugin_name)
-
                 self.bot.unload_extension(
                     f"plugins.{username}-{repo}-{branch}.{name}.{name}"
                 )
-            except Exception:
-                pass
+            except commands.ExtensionNotLoaded:
+                logger.error('Plugin was never loaded.')
 
             self.bot.config["plugins"].remove(plugin_name)
 
             try:
-                # BUG: Local variables 'username', 'branch' and 'repo' might be referenced before assignment
                 if not any(
                     i.startswith(f"{username}/{repo}")
                     for i in self.bot.config["plugins"]
@@ -289,10 +292,9 @@ class Plugins(commands.Cog):
                         f"plugins/{username}-{repo}-{branch}", onerror=onerror
                     )
             except Exception as exc:
-                logger.error(str(exc))
+                logger.error('Failed to remove plugin %s.', plugin_name, exc_info=True)
                 self.bot.config["plugins"].append(plugin_name)
-                logger.error(exc)
-                raise exc
+                return
 
             await self.bot.config.update()
 
@@ -314,9 +316,7 @@ class Plugins(commands.Cog):
 
         if plugin_name in self.registry:
             details = self.registry[plugin_name]
-            plugin_name = (
-                details["repository"] + "/" + plugin_name + "@" + details["branch"]
-            )
+            plugin_name = details["repository"] + "/" + plugin_name + "@" + details["branch"]
 
         if plugin_name not in self.bot.config["plugins"]:
             embed = discord.Embed(
@@ -328,7 +328,8 @@ class Plugins(commands.Cog):
             username, repo, name, branch = self.parse_plugin(plugin_name)
 
             try:
-                cmd = f"cd plugins/{username}-{repo}-{branch} && git reset --hard origin/{branch} && git fetch --all && git pull"
+                cmd = f"cd plugins/{username}-{repo}-{branch} && " \
+                    f"git reset --hard origin/{branch} && git fetch --all && git pull"
                 cmd = await self.bot.loop.run_in_executor(
                     None, self._asubprocess_run, cmd
                 )
@@ -339,6 +340,7 @@ class Plugins(commands.Cog):
                     description=f"An error occurred while updating: {err}.",
                     color=self.bot.main_color,
                 )
+                logger.error('An error occurred while updating plugin:', exc_info=True)
                 await ctx.send(embed=embed)
 
             else:
@@ -357,11 +359,12 @@ class Plugins(commands.Cog):
                     try:
                         await self.load_plugin(username, repo, name, branch)
                     except DownloadError as exc:
-                        em = discord.Embed(
+                        embed = discord.Embed(
                             description=f"Unable to start the plugin: `{exc}`.",
                             color=self.bot.main_color,
                         )
-                        await ctx.send(embed=em)
+                        logger.error('An error occurred while updating plugin:', exc_info=True)
+                        await ctx.send(embed=embed)
 
     @plugin.command(name="enabled", aliases=["installed"])
     @checks.has_permissions(PermissionLevel.OWNER)
@@ -369,7 +372,7 @@ class Plugins(commands.Cog):
         """Shows a list of currently enabled plugins."""
 
         if self.bot.config["plugins"]:
-            msg = "```\n" + "\n".join(self.bot.config["plugins"]) + "\n```"
+            msg = "```\n" + "\n".join(sorted(self.bot.config["plugins"])) + "\n```"
             embed = discord.Embed(description=msg, color=self.bot.main_color)
             await ctx.send(embed=embed)
         else:
@@ -382,19 +385,32 @@ class Plugins(commands.Cog):
         invoke_without_command=True, name="registry", aliases=["list", "info"]
     )
     @checks.has_permissions(PermissionLevel.OWNER)
-    async def plugin_registry(self, ctx, *, plugin_name: str = None):
-        """Shows a list of all approved plugins."""
+    async def plugin_registry(self, ctx, *, plugin_name: typing.Union[int, str] = None):
+        """
+        Shows a list of all approved plugins.
+
+        Usage:
+        `{prefix}plugin registry` Details about all plugins.
+        `{prefix}plugin registry plugin-name` Details about the indicated plugin.
+        `{prefix}plugin registry page-number` Jump to a page in the registry.
+        """
 
         await self.populate_registry()
 
         embeds = []
 
-        registry = list(self.registry.items())
-        random.shuffle(registry)
+        registry = sorted(self.registry.items(), key=lambda elem: elem[0])
 
-        index = next((i for i, (n, _) in enumerate(registry) if plugin_name == n), None)
+        if isinstance(plugin_name, int):
+            index = plugin_name - 1
+            if index < 0:
+                index = 0
+            if index >= len(registry):
+                index = len(registry) - 1
+        else:
+            index = next((i for i, (n, _) in enumerate(registry) if plugin_name == n), 0)
 
-        if index is None:
+        if not index and plugin_name is not None:
             embed = discord.Embed(
                 color=discord.Color.red(),
                 description=f'Could not find a plugin with name "{plugin_name}" within the registry.',
@@ -440,13 +456,13 @@ class Plugins(commands.Cog):
         await paginator.run()
 
     @plugin_registry.command(name="compact")
+    @checks.has_permissions(PermissionLevel.OWNER)
     async def plugin_registry_compact(self, ctx):
         """Shows a compact view of all plugins within the registry."""
 
         await self.populate_registry()
 
-        registry = list(self.registry.items())
-        registry.sort(key=lambda elem: elem[0])
+        registry = sorted(self.registry.items(), key=lambda elem: elem[0])
 
         pages = [""]